+++ quipucords/fingerprinter/runner.py
@@ -35,7 +35,7 @@ from fingerprinter.jboss_fuse import detect_jboss_fuse
 from fingerprinter.jboss_web_server import detect_jboss_ws
 from fingerprinter.utils import strip_suffix
 from scanner.openshift import formatters as ocp_formatters
-from scanner.task import ScanTaskRunner
+from scanner.runner import ScanTaskRunner
 from scanner.vcenter.utils import VcenterRawFacts
 from utils import deepget, default_getter
