+++ quipucords/scanner/runner.py
