+++ quipucords/scanner/satellite/tests_sat_inspect.py
@@ -91,7 +91,7 @@ class InspectTaskRunnerTest(TestCase):
         scan_job, inspect_task = self.create_scan_job()
         task = InspectTaskRunner(scan_job, inspect_task)
         with patch(
-            "scanner.satellite.task.utils.status", return_value=(None, None, None)
+            "scanner.satellite.runner.utils.status", return_value=(None, None, None)
         ) as mock_sat_status:
             status = task.run(Value("i", ScanJob.JOB_RUN))
             mock_sat_status.assert_called_once_with(ANY)
@@ -102,7 +102,7 @@ class InspectTaskRunnerTest(TestCase):
         scan_job, inspect_task = self.create_scan_job()
         task = InspectTaskRunner(scan_job, inspect_task)
         with patch(
-            "scanner.satellite.task.utils.status",
+            "scanner.satellite.runner.utils.status",
             return_value=(401, None, SATELLITE_VERSION_5),
         ) as mock_sat_status:
             status = task.run(Value("i", ScanJob.JOB_RUN))
@@ -115,7 +115,7 @@ class InspectTaskRunnerTest(TestCase):
         task = InspectTaskRunner(scan_job, inspect_task)
 
         with patch(
-            "scanner.satellite.task.utils.status",
+            "scanner.satellite.runner.utils.status",
             return_value=(401, None, SATELLITE_VERSION_6),
         ) as mock_sat_status:
             status = task.run(Value("i", ScanJob.JOB_RUN))
@@ -128,7 +128,7 @@ class InspectTaskRunnerTest(TestCase):
         task = InspectTaskRunner(scan_job, inspect_task)
 
         with patch(
-            "scanner.satellite.task.utils.status",
+            "scanner.satellite.runner.utils.status",
             return_value=(200, 3, SATELLITE_VERSION_6),
         ) as mock_sat_status:
             status = task.run(Value("i", ScanJob.JOB_RUN))
@@ -141,7 +141,7 @@ class InspectTaskRunnerTest(TestCase):
         task = InspectTaskRunner(scan_job, inspect_task)
 
         with patch(
-            "scanner.satellite.task.utils.status", side_effect=mock_conn_exception
+            "scanner.satellite.runner.utils.status", side_effect=mock_conn_exception
         ) as mock_sat_status:
             status = task.run(Value("i", ScanJob.JOB_RUN))
             mock_sat_status.assert_called_once_with(ANY)
@@ -153,7 +153,7 @@ class InspectTaskRunnerTest(TestCase):
         task = InspectTaskRunner(scan_job, inspect_task)
 
         with patch(
-            "scanner.satellite.task.utils.status",
+            "scanner.satellite.runner.utils.status",
             side_effect=mock_sat_auth_exception,
         ) as mock_sat_status:
             status = task.run(Value("i", ScanJob.JOB_RUN))
@@ -166,7 +166,7 @@ class InspectTaskRunnerTest(TestCase):
         task = InspectTaskRunner(scan_job, inspect_task)
 
         with patch(
-            "scanner.satellite.task.utils.status", side_effect=mock_sat_exception
+            "scanner.satellite.runner.utils.status", side_effect=mock_sat_exception
         ) as mock_sat_status:
             status = task.run(Value("i", ScanJob.JOB_RUN))
             mock_sat_status.assert_called_once_with(ANY)
@@ -178,7 +178,7 @@ class InspectTaskRunnerTest(TestCase):
         task = InspectTaskRunner(scan_job, inspect_task)
 
         with patch(
-            "scanner.satellite.task.utils.status", side_effect=mock_timeout_error
+            "scanner.satellite.runner.utils.status", side_effect=mock_timeout_error
         ) as mock_sat_status:
             status = task.run(Value("i", ScanJob.JOB_RUN))
             mock_sat_status.assert_called_once_with(ANY)
@@ -190,7 +190,7 @@ class InspectTaskRunnerTest(TestCase):
         task = InspectTaskRunner(scan_job, inspect_task)
 
         with patch(
-            "scanner.satellite.task.utils.status", side_effect=mock_exception
+            "scanner.satellite.runner.utils.status", side_effect=mock_exception
         ) as mock_sat_status:
             with self.assertRaises(Exception):
                 status = task.run(Value("i", ScanJob.JOB_RUN))
@@ -203,7 +203,7 @@ class InspectTaskRunnerTest(TestCase):
         task = InspectTaskRunner(scan_job, inspect_task)
 
         with patch(
-            "scanner.satellite.task.utils.status",
+            "scanner.satellite.runner.utils.status",
             return_value=(200, 2, SATELLITE_VERSION_6),
         ) as mock_sat_status:
             with patch.object(SatelliteSixV2, "hosts_facts") as mock_facts:
