+++ quipucords/scanner/vcenter/inspect.py
@@ -6,7 +6,7 @@ from django.db import transaction
 from pyVmomi import vim, vmodl  # pylint: disable=no-name-in-module
 
 from api.models import RawFact, ScanTask, SystemInspectionResult
-from scanner.task import ScanTaskRunner
+from scanner.runner import ScanTaskRunner
 from scanner.vcenter.utils import (
     ClusterRawFacts,
     HostRawFacts,
